% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cov.R
\name{dist_cov_creator}
\alias{dist_cov_creator}
\title{Compute Non-Euclidean distances for covariance matrices, according to Dryden AOAS 2009
Useful for subsequently performing hierarchical clustering with different distance types}
\usage{
dist_cov_creator(data, method = "Euclidean")
}
\arguments{
\item{method}{}
}
\description{
Compute Non-Euclidean distances for covariance matrices, according to Dryden AOAS 2009
Useful for subsequently performing hierarchical clustering with different distance types
}
